+++ release-notes.md
@@ -8,6 +8,7 @@ For players:
   * The installer will now automatically clean up old SMAPI files.
   * Each mod now has its own `.cache` folder, so removing the mod won't leave orphaned cache files behind. The installer will automatically remove the old `.cache` folder.
   * Improved installer wording to avoid confusion.
+  * SMAPI will now intercept mod errors in menu draw code, and exit the menu to prevent your game from crashing.
   * Fixed the installer not removing TrainerMod from appdata if it's already in the game mods directory.
   * Fixed the installer not moving mods out of appdata if the game isn't installed on the same Windows partition.
   * Fixed the SMAPI terminal not opening by default on Mac. Linux users are out of luck.